+++ changelog.md
@@ -4,7 +4,9 @@
   ([#235](https://github.com/birtles/rikaichamp/issues/235))
 * Fixed a bug in handling ー after an い sound
   ([#246](https://github.com/birtles/rikaichamp/issues/246))
-  thanks to [@claudiofreitas](https://github.com/claudiofreitas).
+  thanks to [@claudiofreitas](https://github.com/claudiofreitas)
+* Fixed regression to recognizing -masu stem forms
+  ([#245](https://github.com/birtles/rikaichamp/issues/245))
 
 ## 0.1.17 (2020-04-31)